+++ packages/app/postcss.config.cjs
@@ -1,7 +1,6 @@
 module.exports = {
     plugins: {
-      tailwindcss: {},
-      autoprefixer: {}
-    }
-  };
-  
+        tailwindcss: {},
+        autoprefixer: {},
+    },
+}
