+++ packages/app/src/plugin-host/register.ts
