+++ packages/app/src/plugin-worker/register.ts
