+++ packages/app/src/plugin-worker/service.ts
