+++ packages/app/src/styles/tailwind.css
@@ -5,6 +5,5 @@ body {
     box-sizing: border-box;
 }
 
-
 @tailwind components;
 @tailwind utilities;
