+++ packages/app/tsconfig.worker.json
@@ -1,11 +1,11 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "rootDir": "./plugin-worker",
+    "rootDir": "src/plugin-worker",
     "outDir": "./dist/plugin-worker",
     "tsBuildInfoFile": "./dist/plugin-worker.tsbuildinfo",
     "lib": ["WebWorker", "ES2022"]
   },
-  "include": ["./plugin-worker"],
+  "include": ["src/plugin-worker"],
   "references": [{ "path": "../base" }, { "path": "../shared-base" }, { "path": "../plugin-infra" }]
 }
